Initial commit
This is just an import of my helper scripts for grabbing packages from
the AUR, for cases where an AUR helper isn't desirable.

The scripts themselves are as simple as can be: they simply clone
packages into a directory, view PKGBUILDs with a pager, and build with
makepkg -- pretty much just automating what I've been doing to install
AUR packages.

The aursearch script is pretty simple as well: it makes use of the AUR's
HTTP API, and pipes the result to jq to make human-readable.