Add color output to aursearch
To make aursearch results easier on the eyes to scan, the output has
been colored in a similar manner to how pacman -Ss does it (with colors
enabled, of course).

Since piping into less seems to crop up as a usual use case for me, care
has also been taken to forgo color output if stdout is not a tty.