Add aurquery script
Having aursearch to mimic pacman -Ss for the AUR is very convenient, but
I still sometimes find I have to open the web interface to check
additional details for a package (or otherwise fetch the PKGBUILD).

Since the AUR RPC interface can already provide additional details per
package (in the same vein as pacman -Si), I made a script to fetch and
display this info. The formatting code is a bit hack-ish at times, but
at least it mimics pacman -Si input for familiarity.